api/requirements.txt
tenserflow
fastapi
uvicorn
python-multipart
pillow
tensorflow-serving-api
matplotlib
numpy